refactor: 원시 타입을 포장함으로써 유효성 검사 분리
